requirements.txt
google-cloud-storage==2.11.0
google-cloud-bigquery==3.11.4
google-cloud-pubsub==2.18.4
google-cloud-dataproc==5.6.0
google-cloud-vision==3.4.4
google-cloud-language==2.11.1
google-api-python-client==2.101.0
google-auth==2.23.2
pytest==7.4.2
pylint==3.0.0
pytest-cov==4.1.0
black==23.9.1
pyspark==3.5.0